+++ CleanArchitectureTests/core/airquality/AirQualityWorkerTest.swift
@@ -39,10 +39,9 @@ class AirQualityWorkerTest: XCTestCase {
     }
 
     func testGivenSuccessResponseWhenFetchThenMatchExpected() throws {
+        let response = try OHHTTPStubsResponse._200(jsonFileName: "airquality-success.json", inBundleForClass: type(of: self))
         stub(condition: isMethodGET() && isHost(host) && isPath(path) && containsQueryParams(params)) { (request) -> OHHTTPStubsResponse in
-            return OHHTTPStubsResponse(fileAtPath: OHPathForFile("airquality-success.json", type(of: self))!,
-                                       statusCode: 200,
-                                       headers: ["Content-Type":"application/json"])
+            return response
         }.name = "air quality success request"
         
         var result: AirQualityData?
@@ -80,10 +79,7 @@ class AirQualityWorkerTest: XCTestCase {
     
     func testGiven400WhenFetchThenMatchError() throws {
         stub(condition: isMethodGET() && isHost(host) && isPath(path) && containsQueryParams(params)) { (request) -> OHHTTPStubsResponse in
-            let BadRequest = NSError(domain: NSURLErrorDomain,
-                                     code: 400,
-                                     userInfo: ["NSLocalizedDescription": "Bad Request"])
-            return OHHTTPStubsResponse(error: BadRequest)
+            return OHHTTPStubsResponse._400()
         }.name = "air quality no network request"
         
         var result: AirQualityError?
